refactor: integrated box packing/plotting into the app package
