+++ app/plot.py
@@ -1,44 +1,9 @@
-import math
-from itertools import permutations
-
-import py3dbp
 import matplotlib.pyplot as plt
 from mpl_toolkits.mplot3d.art3d import Poly3DCollection
+from py3dbp import Packer
 
 
-def create_containers(max_dim: int = 12) -> list:
-    """Creates containers for all permutations of dimensions.
-
-    The maximum dimension for any side is set by max_dim.
-    """
-    num_list = []
-    for i in range(1, max_dim + 1):
-        num_list += [i]*3
-    container_dims = list(set(permutations(num_list, 3)))
-    container_vol_list = [math.prod(container_dim)
-                          for container_dim
-                          in container_dims]
-    container_dims_sort = [cont
-                           for vol, cont
-                           in sorted(zip(container_vol_list, container_dims))]
-    container_list = [py3dbp.Bin(str(i), dim[0], dim[1], dim[2], 100)
-                      for i, dim
-                      in enumerate(container_dims_sort)]
-    return container_list
-
-
-def pack_items(container_box: py3dbp.Bin,
-               items: list[py3dbp.Item]) -> py3dbp.Packer:
-    """Attempts to pack items into container"""
-    packer = py3dbp.Packer()
-    packer.add_bin(container_box)
-    for item in items:
-        packer.add_item(item)
-    packer.pack(number_of_decimals=0)
-    return container_box
-
-
-def plot_result(result: py3dbp.Packer) -> None:
+def plot_result(result: Packer) -> None:
     """Plots a solution. Limited to 10 items by COLORS_LIST"""
     width = float(result.width)
     height = float(result.height)
@@ -183,34 +148,3 @@ def plot_result(result: py3dbp.Packer) -> None:
     fig.show()
 
     input("press to continue")
-
-
-# A small test to see if this works or not
-if __name__ == '__main__':
-    container_list = create_containers(max_dim=5)
-
-    item_list = [py3dbp.Item('Box-1', 3, 1, 2, 1),
-                 py3dbp.Item('Box-2', 3, 1, 2, 1),
-                 py3dbp.Item('Box-3', 3, 1, 2, 1),
-                 py3dbp.Item('Box-4', 3, 1, 2, 1),
-                 py3dbp.Item('Box-5', 3, 1, 2, 1)]
-
-    for container in container_list:
-        result = pack_items(container, item_list)
-        if not result.unfitted_items:
-            print("All items fit!")
-            print(" ")
-            break
-        else:
-            result = None
-
-    if not result:
-        print("There was no fit!")
-    else:
-        print(f'''Container Dimensions
-    Width: {result.width}
-    Height: {result.height}
-    Depth: {result.depth}
-    ''')
-
-    plot_result(result)
