+++ test_app.py
@@ -1,6 +1,6 @@
 from blessed import Terminal
 
-from app import InsideTheBoxTUI
+from app.app import InsideTheBoxTUI
 
 if __name__ == '__main__':
     InsideTheBoxTUI(Terminal()).run()
